pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lightweight-charts-pro-backend"
version = "0.1.0"
description = "FastAPI backend for TradingView Lightweight Charts Pro"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Streamlit Lightweight Charts Contributors"}
]
keywords = [
    "lightweight-charts",
    "tradingview",
    "charts",
    "fastapi",
    "websocket",
    "backend"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
    "slowapi>=0.1.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "python-json-logger>=2.0.0",
    "prometheus-fastapi-instrumentator>=6.1.0",
    # Note: lightweight-charts-pro is an optional dependency for local development
    # It's not required for documentation builds or basic API functionality
    # "lightweight-charts-pro @ file:///Users/nandkapadia/lightweight-charts-pro-python",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pytest-cov>=4.0.0",
    "isort>=5.12.0",
    "autoflake>=2.2.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "bandit>=1.7.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=2.0.0",
    "sphinx-autobuild>=2021.3.14",
]

[project.urls]
Homepage = "https://github.com/nandkapadia/streamlit-lightweight-charts-pro"
Documentation = "https://github.com/nandkapadia/streamlit-lightweight-charts-pro#readme"
Repository = "https://github.com/nandkapadia/streamlit-lightweight-charts-pro.git"

[tool.setuptools.packages.find]
where = ["."]
include = ["lightweight_charts_pro_backend*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
float_to_top = true
skip_gitignore = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.autoflake]
remove-all-unused-imports = true
remove-unused-variables = true
in-place = true
